+++ mne/viz/backends/_abstract.py
@@ -5,7 +5,9 @@
 #
 # License: Simplified BSD
 
+import warnings
 from abc import ABC, abstractmethod, abstractclassmethod
+from ..utils import tight_layout
 
 
 class _AbstractRenderer(ABC):
@@ -586,6 +588,84 @@ class _AbstractWidget(ABC):
         pass
 
 
+class _AbstractMplCanvas(ABC):
+    @abstractmethod
+    def __init__(self, brain, width, height, dpi):
+        """Initialize the MplCanvas."""
+        pass
+
+    def plot(self, x, y, label, **kwargs):
+        """Plot a curve."""
+        line, = self.axes.plot(
+            x, y, label=label, **kwargs)
+        self.update_plot()
+        return line
+
+    def plot_time_line(self, x, label, **kwargs):
+        """Plot the vertical line."""
+        line = self.axes.axvline(x, label=label, **kwargs)
+        self.update_plot()
+        return line
+
+    def update_plot(self):
+        """Update the plot."""
+        leg = self.axes.legend(
+            prop={'family': 'monospace', 'size': 'small'},
+            framealpha=0.5, handlelength=1.,
+            facecolor=self.brain._bg_color)
+        for text in leg.get_texts():
+            text.set_color(self.brain._fg_color)
+        with warnings.catch_warnings(record=True):
+            warnings.filterwarnings('ignore', 'constrained_layout')
+            self.canvas.draw()
+
+    def set_color(self, bg_color, fg_color):
+        """Set the widget colors."""
+        self.axes.set_facecolor(bg_color)
+        self.axes.xaxis.label.set_color(fg_color)
+        self.axes.yaxis.label.set_color(fg_color)
+        self.axes.spines['top'].set_color(fg_color)
+        self.axes.spines['bottom'].set_color(fg_color)
+        self.axes.spines['left'].set_color(fg_color)
+        self.axes.spines['right'].set_color(fg_color)
+        self.axes.tick_params(axis='x', colors=fg_color)
+        self.axes.tick_params(axis='y', colors=fg_color)
+        self.fig.patch.set_facecolor(bg_color)
+
+    @abstractmethod
+    def show(self):
+        """Show the canvas."""
+        pass
+
+    def close(self):
+        """Close the canvas."""
+        self.canvas.close()
+
+    def on_button_press(self, event):
+        """Handle button presses."""
+        # left click (and maybe drag) in progress in axes
+        if (event.inaxes != self.axes or
+                event.button != 1):
+            return
+        self.time_func(
+            event.xdata, update_widget=True, time_as_index=False)
+
+    def clear(self):
+        """Clear internal variables."""
+        self.close()
+        self.axes.clear()
+        self.fig.clear()
+        self.brain = None
+        self.canvas = None
+        self.manager = None
+
+    on_motion_notify = on_button_press  # for now they can be the same
+
+    def on_resize(self, event):
+        """Handle resize events."""
+        tight_layout(fig=self.axes.figure)
+
+
 class _AbstractWindow(ABC):
     @abstractmethod
     def _window_initialize(self, func=None):
@@ -597,6 +677,20 @@ class _AbstractWindow(ABC):
 
     @abstractmethod
     def _window_get_size(self):
+        pass
+
+    def _window_get_mplcanvas_size(self, ratio):
+        dpi = self._window_get_dpi()
+        w, h = self._window_get_size()
+        h /= ratio
+        return (w / dpi, h / dpi)
+
+    @abstractmethod
+    def _window_get_mplcanvas(self, brain, ratio):
+        pass
+
+    @abstractmethod
+    def _window_adjust_mplcanvas_layout(self):
         pass
 
     @abstractmethod
